Split run records and the run lock out of isolation-state
isolation-state.ts had grown past the 400-line target. Reading and
writing a run's record now lives in isolation-records.ts, and the lock a
run holds over its folder lives in isolation-lock.ts. Every file in the
mutation isolation set is now under 300 lines.

The name of the half-written record file is no longer exported: only the
writer uses it.
