+++ scripts/mutation/isolation-cleanup.ts
@@ -8,13 +8,12 @@
 import { rethrowUnlessNotFound } from "#scripts/not-found.ts";
 import { processExists, removeTree } from "#scripts/process.ts";
 import { errorMessage } from "#shared/error-message.ts";
+import { runLockIsHeld } from "./isolation-lock.ts";
+import { readRunRecords, runDirectoryNames } from "./isolation-records.ts";
 import {
   MUTATION_RUN_ID_PREFIX,
   type MutationRunRecord,
   newRunRecord,
-  readRunRecords,
-  runDirectoryNames,
-  runLockIsHeld,
   runRoot,
   runStartedRecently,
   withinStartupGrace,
