+++ test/scripts/mutation/isolation/list-and-kill.test.ts
@@ -1,13 +1,13 @@
 import { expect } from "@std/expect";
 import { describe, it as test } from "@std/testing/bdd";
 import { stub } from "@std/testing/mock";
+import { withMutationRunLock } from "#scripts/mutation/isolation-lock.ts";
+import { writeRunRecord } from "#scripts/mutation/isolation-records.ts";
 import {
   ISOLATION_USAGE,
   markFinished,
   markRunning,
   newRunRecord,
-  withMutationRunLock,
-  writeRunRecord,
 } from "#scripts/mutation/isolation-state.ts";
 import {
   captureMutationCommand,
